Temperatuur sensor volledig functioneel
Temperatuur sensor werkt nu volledig met een 10 bits precisie en is geintegreerd met de code voor de ultrasoonsensor. Led lampje blink is veranderd van sleep naar een geschedulde task.

+++ C/Embedded/embedded.h
@@ -20,4 +20,4 @@ void set_screenstate(bool state);
 
 // Function prototypes voor de ADC_handler
 void ADC_init(void);
-uint8_t get_ADCValue();
+uint16_t get_ADCValue();
